feat: update popular token cache
