+++ packages/espressocash_app/lib/features/popular_tokens/data/coingecko_client.dart
@@ -5,6 +5,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
 import 'package:retrofit/retrofit.dart';
 
+import '../../../core/coingecko_client.dart';
 import '../../../core/tokens/token.dart';
 import '../../../core/tokens/token_list.dart';
 
@@ -15,9 +16,11 @@ part 'coingecko_client.g.dart';
 @RestApi(baseUrl: 'https://api.coingecko.com/api/v3')
 abstract class MarketsCoingeckoClient {
   @factoryMethod
-  factory MarketsCoingeckoClient(Dio dio) = _MarketsCoingeckoClient;
+  factory MarketsCoingeckoClient(CoingeckoClient client) =>
+      _MarketsCoingeckoClient(client.dio);
 
   @GET('/coins/markets')
+  @Extra({maxAgeOption: Duration(hours: 1)})
   Future<List<MarketsResponseDto>> getMarketTokens(
     @Queries() MarketsRequestDto request,
   );
