+++ packages/espressocash_app/lib/features/popular_tokens/data/repository.dart
@@ -6,29 +6,17 @@ import 'package:injectable/injectable.dart';
 import '../../../core/tokens/token.dart';
 import '../../../core/tokens/token_list.dart';
 import 'coingecko_client.dart';
-import 'popular_token_cache.dart';
 
 @injectable
 class PopularTokenRepository {
   PopularTokenRepository({
     required MarketsCoingeckoClient coingeckoClient,
-    required PopularTokenCache popularTokenCache,
     required TokenList tokenList,
   })  : _coingeckoClient = coingeckoClient,
-        _cache = popularTokenCache,
         _tokenList = tokenList;
 
   final MarketsCoingeckoClient _coingeckoClient;
-  final PopularTokenCache _cache;
   final TokenList _tokenList;
-
-  AsyncResult<IMap<Token, double>> get({required String currency}) async {
-    final cachedResult = await _cache.get();
-
-    return cachedResult.isNotEmpty
-        ? Either.right(cachedResult)
-        : await refresh(currency: currency);
-  }
 
   AsyncResult<IMap<Token, double>> refresh({required String currency}) =>
       _coingeckoClient
@@ -48,8 +36,7 @@ class PopularTokenRepository {
                   .let((t) => MapEntry(t, r.currentPrice ?? 0)),
             ),
           )
-          .mapAsync(IMap.fromEntries)
-          .doOnRightAsync(_cache.set);
+          .mapAsync(IMap.fromEntries);
 
   List<MarketsResponseDto> _sortTokens(
     List<MarketsResponseDto> response,
@@ -63,8 +50,6 @@ class PopularTokenRepository {
       ..remove(sol)
       ..insert(0, sol);
   }
-
-  Future<void> clear() => _cache.remove();
 }
 
 const _popularTokensCoinGeckoId = [
